+++ pkg/database/dgraph.go
@@ -6,6 +6,8 @@ import (
 	"errors"
 	"fmt"
 	"log"
+	"reflect"
+	"strings"
 	"sync"
 	"sync/atomic"
 	"text/template"
@@ -25,9 +27,13 @@ path: string @index(trigram) .
 hash: string @index(exact) .
 type: string @index(hash) .
 name: string @index(hash) .
-nodeType: int @index(int) .
 dataNodes: uid @reverse .
 data: string @index(hash) .
+packageNodeType: string @index(hash) .
+fileNodeType: string @index(hash) .
+infoNodeType: string @index(hash) .
+dataNodeType: string @index(hash) .
+analyzerNodeType: string @index(hash) .
 `
 
 type DataBase struct {
@@ -89,15 +95,6 @@ func (db *DataBase) AwaitBuildComplete() {
 	close(db.insertQueue)
 }
 
-// AddFileNode adds a node to the insert queue
-func (db *DataBase) AddFileNode(node *service.FileNode) {
-	atomic.AddUint64(&db.pending, 1)
-	for _, dep := range node.DerivedFrom {
-		db.AddFileNode(dep)
-	}
-	db.insertQueue <- node
-}
-
 // the queueWorker runs in a go routine and inserts the nodes from the insert queue into the database
 func queueWorker(db *DataBase) {
 	for {
@@ -137,7 +134,6 @@ func queueWorker(db *DataBase) {
 		if uid != "" {
 			node.Uid = uid
 		}
-		service.SanitizeFileNode(node)
 		uid, err = dbInsert(db.client, node)
 		if err != nil {
 			panic(err)
@@ -147,298 +143,50 @@ func queueWorker(db *DataBase) {
 	}
 }
 
-func (db *DataBase) GetAllInfoData(infotype string) ([]string, error) {
-	const q = `
-	query InfoData($itype: string){
-		getInfoData(func:eq(nodeType, 2))  @filter(eq(type, $itype)) {
-			A as dataNodes
+func fixTypeField(field *reflect.Value) error {
+	switch field.Kind() {
+	case reflect.Struct:
+		for i := 0; i < field.NumField(); i++ {
+			fieldVal := field.Field(i)
+			if fieldVal.Kind() == reflect.String && strings.HasSuffix(field.Type().Field(i).Name, "NodeType") {
+				if fieldVal.CanSet() {
+					fieldVal.SetString("_")
+					continue
+				}
+				return fmt.Errorf("%s not settable", field.Type().Field(i).Name)
+			}
+			fixTypeField(&fieldVal)
 		}
-
-		infodata(func: uid(A)) {
-			data
-		}
-	}
-	`
-	vars := map[string]string{"$itype": infotype}
-
-	resp, err := db.client.NewTxn().QueryWithVars(context.Background(), q, vars)
-	if err != nil {
-		return nil, err
-	}
-
-	type Data struct {
-		Data string
-	}
-
-	type InfoData struct {
-		Infodata []Data
-	}
-
-	var r InfoData
-
-	err = json.Unmarshal(resp.Json, &r)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	ret := []string{}
-	for _, data := range r.Infodata {
-		ret = append(ret, data.Data)
-	}
-	return ret, nil
-}
-
-func (db *DataBase) GetInfoData(rootNodeID string, infotype string, datatype string) ([]string, error) {
-	const q = `
-	query InfoData($id: string, $itype: string, $dtype: string){
-		var(func: uid($id)) @recurse(loop: false) {
-			uid
-			nodeType
-			name:type
-			analyzer
-			trustLevel
-			A as dataNodes @filter(eq(type, $dtype))
-			data
-			confidenceScore
-			additionalInfo @filter(eq(type, $itype)) (orderdesc: confidenceScore, first: 1)
-			derivedFrom
-		}
-
-		infodata(func: uid(A)) {
-			data
+	case reflect.Slice, reflect.Array:
+		for i := 0; i < field.Len(); i++ {
+			fieldVal := field.Index(i).Elem()
+			fixTypeField(&fieldVal)
 		}
 	}
-	`
-
-	vars := map[string]string{"$id": rootNodeID, "$itype": infotype, "$dtype": datatype}
-
-	resp, err := db.client.NewTxn().QueryWithVars(context.Background(), q, vars)
-	if err != nil {
-		return nil, err
-	}
-
-	type Data struct {
-		Data string
-	}
-
-	type InfoData struct {
-		Infodata []Data
-	}
-
-	var r InfoData
-
-	err = json.Unmarshal(resp.Json, &r)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	ret := []string{}
-	for _, data := range r.Infodata {
-		ret = append(ret, data.Data)
-	}
-	return ret, nil
-}
-
-func (db *DataBase) AddInfoNodes(nodeID string, infonodes ...*service.InfoNode) error {
-	db.insertMutex.Lock()
-	defer db.insertMutex.Unlock()
-
-	const q = `
-	query Node($id: string){
-		node(func: uid($id)) @recurse(loop: false) {
-			uid
-			nodeType
-			additionalInfo
-		}
-	}
-	`
-	vars := map[string]string{"$id": nodeID}
-	resp, err := db.client.NewTxn().QueryWithVars(context.Background(), q, vars)
-	if err != nil {
-		return err
-	}
-
-	type GenericNode struct {
-		Uid            string
-		NodeType       int
-		AdditionalInfo []*service.InfoNode
-	}
-
-	type GenericRoot struct {
-		Node []GenericNode
-	}
-	var r GenericRoot
-
-	err = json.Unmarshal(resp.Json, &r)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if len(r.Node) < 1 {
-		return fmt.Errorf("No node with uid %s found", nodeID)
-	}
-
-	receiverNode := r.Node[0]
-
-	switch receiverNode.NodeType {
-	case service.NodeTypeFileNode:
-		type FileRoot struct {
-			Node []service.FileNode
-		}
-		var r FileRoot
-
-		err = json.Unmarshal(resp.Json, &r)
-		if err != nil {
-			log.Fatal(err)
-		}
-		fileNode := r.Node[0]
-		fileNode.AdditionalInfo = append(fileNode.AdditionalInfo, infonodes...)
-		service.SanitizeFileNode(&fileNode)
-		_, err = dbInsert(db.client, fileNode)
-		if err != nil {
-			return err
-		}
-	case service.NodeTypePackageNode:
-		type PackageRoot struct {
-			Node []service.PackageNode
-		}
-		var r PackageRoot
-
-		err = json.Unmarshal(resp.Json, &r)
-		if err != nil {
-			log.Fatal(err)
-		}
-		pkgNode := r.Node[0]
-		pkgNode.AdditionalInfo = append(pkgNode.AdditionalInfo, infonodes...)
-		service.SanitizePackageNode(&pkgNode)
-		log.Printf("Package node uid %s", pkgNode.Uid)
-		_, err := dbInsert(db.client, pkgNode)
-		if err != nil {
-			return err
-		}
-	default:
-		return errors.New("can not attach infonode, receiver is neither file nor package node")
-	}
-
 	return nil
 }
 
-func (db *DataBase) AddFileNodes(nodeID string, filenodes ...*service.FileNode) error {
-	db.insertMutex.Lock()
-	defer db.insertMutex.Unlock()
-
-	const q = `
-	query Node($id: string){
-		node(func: uid($id)) @recurse(loop: false) {
-			uid
-			nodeType
-			targets
-			derivedFrom
-		}
+// fill in the type value to work around omitting empty values on serialization
+func fillTypeField(data interface{}) error {
+	val := reflect.ValueOf(data)
+	if val.Kind() != reflect.Ptr {
+		return errors.New("you shall not call fillTypeField by value")
 	}
-	`
-	vars := map[string]string{"$id": nodeID}
-	resp, err := db.client.NewTxn().QueryWithVars(context.Background(), q, vars)
-	if err != nil {
-		return err
+	for val.Kind() == reflect.Ptr || val.Kind() == reflect.Interface {
+		val = val.Elem()
 	}
 
-	type GenericNode struct {
-		Uid      string
-		NodeType int
-	}
-
-	type GenericRoot struct {
-		Node []GenericNode
-	}
-	var r GenericRoot
-
-	err = json.Unmarshal(resp.Json, &r)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if len(r.Node) < 1 {
-		return fmt.Errorf("No node with uid %s found", nodeID)
-	}
-	receiverNode := r.Node[0]
-
-	switch receiverNode.NodeType {
-	case service.NodeTypeFileNode:
-		type FileRoot struct {
-			Node []service.FileNode
-		}
-		var r FileRoot
-
-		err = json.Unmarshal(resp.Json, &r)
-		if err != nil {
-			log.Fatal(err)
-		}
-		fileNode := r.Node[0]
-		fileNode.DerivedFrom = append(fileNode.DerivedFrom, filenodes...)
-		service.SanitizeFileNode(&fileNode)
-		_, err = dbInsert(db.client, fileNode)
-		if err != nil {
-			return err
-		}
-	case service.NodeTypePackageNode:
-		type PackageRoot struct {
-			Node []service.PackageNode
-		}
-		var r PackageRoot
-
-		err = json.Unmarshal(resp.Json, &r)
-		if err != nil {
-			log.Fatal(err)
-		}
-		pkgNode := r.Node[0]
-		pkgNode.Targets = append(pkgNode.Targets, filenodes...)
-		service.SanitizePackageNode(&pkgNode)
-		log.Printf("Package node uid %s", pkgNode.Uid)
-		_, err := dbInsert(db.client, pkgNode)
-		if err != nil {
-			return err
-		}
-	default:
-		return errors.New("can not attach infonode, receiver is neither file nor package node")
-	}
-
-	return nil
-}
-
-// GetFileNodeUid returns the UID of the node if exists otherwise ""
-func (db *DataBase) GetFileNodeUid(hash string) (string, error) {
-
-	ret := map[string][]*service.FileNode{}
-
-	q := `query Node($Hash: string){
-		  hasNode(func: eq(hash, $Hash)) {
-			uid
-		  }}`
-
-	vars := map[string]string{"$Hash": hash}
-
-	err := db.queryNodes(q, vars, &ret)
-	if err != nil {
-		return "", err
-	}
-
-	// no node with such hash
-	if len(ret["hasNode"]) == 0 {
-		return "", nil
-	}
-	return ret["hasNode"][0].Uid, nil
-}
-
-func (db *DataBase) AddPackageNode(pNode *service.PackageNode) (string, error) {
-	pNode.NodeType = service.NodeTypePackageNode
-	return dbInsert(db.client, pNode)
+	return fixTypeField(&val)
 }
 
 // the data should be JSON marshalable
 func dbInsert(c *client.Dgraph, data interface{}) (string, error) {
 	txn := c.NewTxn()
 	defer txn.Discard(context.Background())
+
+	if err := fillTypeField(&data); err != nil {
+		return "", err
+	}
 
 	jNode, err := json.Marshal(data)
 	if err != nil {
@@ -462,226 +210,6 @@ func getVarName(index int) string {
 	}
 	result = result + string(rune(65+index))
 	return result
-}
-
-func (db *DataBase) GetInfoNodeByDataNode(infonodetype string, datanodes ...*service.InfoNode_DataNode) (*service.InfoNode, error) {
-
-	var retInfoNode *service.InfoNode
-
-	runeDataNodeMap := map[string]*service.InfoNode_DataNode{}
-
-	for idx, datanode := range datanodes {
-		service.SanitizeDataNode(datanodes[idx])
-		runeDataNodeMap[getVarName(idx)] = datanode
-	}
-
-	ret := map[string][]*service.InfoNode{}
-
-	q := `query InfoNodeByDataNode($InfoType: string) {
-				{{range $var, $data := .}}
-				var(func:eq(nodeType, 3)) @filter(eq(type, "{{$data.Type}}") AND eq(data, "{{$data.Data}}")) {
-					{{$var}} as ~dataNodes
-				}
-				{{end}}
-		
-				getInfoByData(func:eq(nodeType, 2)) @filter(eq(type, $InfoType) {{range $var, $data := .}} AND uid({{$var}}) {{end}}) {
-					uid
-					type
-					dataNodes {
-						type
-						data
-					}
-				}
-			  }`
-
-	queryTmpl, err := template.New("infobydata").Parse(q)
-
-	var b bytes.Buffer
-	err = queryTmpl.Execute(&b, runeDataNodeMap)
-	if err != nil {
-		panic(err)
-	}
-
-	queryString := b.String()
-
-	vars := map[string]string{"$InfoType": infonodetype}
-
-	err = db.queryInfoNodes(queryString, vars, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(ret["getInfoByData"]) > 0 {
-		retInfoNode = ret["getInfoByData"][0]
-	}
-
-	if retInfoNode == nil {
-		infoNode := &service.InfoNode{Type: infonodetype, NodeType: service.NodeTypeInfoNode}
-		infoNode.DataNodes = datanodes
-		uid, err := dbInsert(db.client, infoNode)
-		if err != nil {
-			return nil, err
-		}
-		infoNode.Uid = uid
-		retInfoNode = infoNode
-	}
-
-	return retInfoNode, nil
-}
-
-func (db *DataBase) GetPackageNode(session string) (*service.PackageNode, error) {
-	ret := map[string][]*service.PackageNode{}
-
-	q := `query PackageNode($Session: string) {
-		getPackageNode(func: eq(nodeType, 5)) @recurse(loop: false) {
-			uid
-			session
-			buildConfig
-			nodeType
-			hash
-			name
-			type
-			targets
-			derivedFrom
-			path
-			additionalInfo
-			dataNodes
-			data
-		  }}`
-
-	vars := map[string]string{"$Session": session}
-
-	err := db.queryPackage(q, vars, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(ret["getPackageNode"]) < 1 {
-		return nil, errors.New("No package node found")
-	}
-
-	return ret["getPackageNode"][0], nil
-
-}
-
-func (db *DataBase) GetFileNodesByType(filetype string, recursive bool) ([]*service.FileNode, error) {
-	ret := map[string][]*service.FileNode{}
-
-	q := `query FileNodeByType($FileType: string){
-		  getFileNodeByType(func: eq(nodeType, 1)) {{.Filter}} {{.Recurse}}{
-			uid
-			hash
-			path
-			derivedFrom
-		  }}`
-
-	queryTmpl, err := template.New("filenodesbytype").Parse(q)
-
-	type QueryParams struct {
-		Recurse string
-		Filter  string
-	}
-
-	qp := QueryParams{}
-	if recursive {
-		qp.Recurse = "@recurse(loop: false)"
-	}
-	if filetype != "" {
-		qp.Filter = "@filter(eq(type, $FileType))"
-	}
-
-	var b bytes.Buffer
-	err = queryTmpl.Execute(&b, qp)
-	if err != nil {
-		panic(err)
-	}
-
-	vars := map[string]string{"$FileType": filetype}
-
-	err = db.queryNodes(b.String(), vars, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	return ret["getFileNodeByType"], nil
-}
-
-func (db *DataBase) GetFileNodesByFileNode(filenode *service.FileNode, recursive bool) ([]*service.FileNode, error) {
-	ret := map[string][]*service.FileNode{}
-
-	q := `query FileNodeByFileNode($Filter: string){
-		getFileNodeByFileNode(func: eq(nodeType, 1)) {{.Type}} {{.Recurse}}{
-		  uid
-		  hash
-		  path
-		  derivedFrom
-		}}`
-
-	queryTmpl, err := template.New("filenodesbyfilenode").Parse(q)
-
-	type QueryParams struct {
-		Recurse string
-		Type    string
-		Filter  string
-	}
-
-	qp := QueryParams{}
-
-	if recursive {
-		qp.Recurse = "@recurse(loop: false)"
-	}
-	if filenode.Type != "" {
-		qp.Type = "@filter(eq(type, $Filter))"
-		qp.Filter = filenode.Type
-	}
-
-	var b bytes.Buffer
-	err = queryTmpl.Execute(&b, qp)
-	if err != nil {
-		panic(err)
-	}
-
-	vars := map[string]string{"$Filter": qp.Filter}
-
-	err = db.queryNodes(b.String(), vars, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	return ret["getFileNodeByFileNode"], nil
-
-}
-
-func (db *DataBase) GetAnalyzerByName(name string) (*service.Analyzer, error) {
-	ret := map[string][]*service.Analyzer{}
-
-	q := `query AnalyzerByName($AnaName: string){
-		  getAnalyzerByType(func: eq(nodeType, 4)) @filter(eq(name, $AnaName)) {
-			uid
-			hash
-			path
-			derivedFrom
-		  }}`
-
-	vars := map[string]string{"$AnaName": name}
-
-	err := db.queryAnalyzer(q, vars, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(ret["getAnalyzerByName"]) < 1 {
-		// No such analyzer
-		analyzer := &service.Analyzer{Name: name, NodeType: service.NodeTypeAnalyzerNode}
-		uid, err := dbInsert(db.client, analyzer)
-		if err != nil {
-			return nil, err
-		}
-		analyzer.Uid = uid
-		return analyzer, nil
-	}
-
-	return ret["getAnalyzerByName"][0], nil
 }
 
 func (db *DataBase) GetNodesByType(valuetype string, recursive bool, namefilter string) ([]*service.FileNode, error) {
@@ -727,37 +255,6 @@ func (db *DataBase) GetNodesByType(valuetype string, recursive bool, namefilter 
 	return ret["getNodeByType"], nil
 }
 
-func (db *DataBase) GetNodeByHash(hash string, recursive bool) (*service.FileNode, error) {
-
-	ret := map[string][]*service.FileNode{}
-
-	q := `query NodeByHash($Hash: string){
-		  getNodeByHash(func: eq(hash, $Hash)) {
-			uid
-			hash
-			path
-		  }}`
-
-	if recursive {
-		q = `query NodeByHash($Hash: string){
-		  getNodeByHash(func: eq(hash, $Hash)) @recurse(loop: false) {
-			uid
-			hash
-			path
-			derivedFrom
-		  }}`
-	}
-
-	vars := map[string]string{"$Hash": hash}
-
-	err := db.queryNodes(q, vars, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	return ret["getNodeByHash"][0], nil
-}
-
 func (db *DataBase) queryPackage(query string, queryVars map[string]string, resultMap *map[string][]*service.PackageNode) error {
 	resp, err := db.client.NewTxn().QueryWithVars(context.Background(), query, queryVars)
 	if err != nil {
@@ -774,6 +271,18 @@ func (db *DataBase) queryNodes(query string, queryVars map[string]string, result
 	resp, err := db.client.NewTxn().QueryWithVars(context.Background(), query, queryVars)
 	if err != nil {
 		return fmt.Errorf("Could not query for filenode with: \n\n%s\n\nVars:\n\n%v\n\nError: %v", query, queryVars, err)
+	}
+
+	if err = json.Unmarshal(resp.Json, resultMap); err != nil {
+		return fmt.Errorf("Could not unmashal query response: %v", err)
+	}
+	return nil
+}
+
+func (db *DataBase) queryAnyNodes(query string, queryVars map[string]string, resultMap *map[string][]interface{}) error {
+	resp, err := db.client.NewTxn().QueryWithVars(context.Background(), query, queryVars)
+	if err != nil {
+		return fmt.Errorf("Could not query for node with: \n\n%s\n\nVars:\n\n%v\n\nError: %v", query, queryVars, err)
 	}
 
 	if err = json.Unmarshal(resp.Json, resultMap); err != nil {
