+++ pkg/database/dgraph_test.go
@@ -1,6 +1,10 @@
 package database
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/QMSTR/qmstr/pkg/service"
+)
 
 func TestVarNameCalculation(t *testing.T) {
 	if getVarName(0) != "A" {
@@ -15,4 +19,16 @@ func TestVarNameCalculation(t *testing.T) {
 	if getVarName(52) != "AAA" {
 		t.Fail()
 	}
+}
+
+func TestFillType(t *testing.T) {
+	pkgNode := service.PackageNode{Targets: []*service.FileNode{&service.FileNode{Name: "Test"}}}
+	fillTypeField(&pkgNode)
+	if pkgNode.PackageNodeType != "_" {
+		t.Fail()
+	}
+	if pkgNode.Targets[0].FileNodeType != "_" {
+		t.Fail()
+	}
+
 }
