+++ pkg/master/init.go
@@ -30,9 +30,9 @@ func (phase *serverPhaseInit) Activate() error {
 
 func (phase *serverPhaseInit) initPackage(session string) error {
 	rootPackageNode := &service.PackageNode{Name: phase.masterConfig.Name, BuildConfig: phase.masterConfig.BuildConfig}
-	tmpInfoNode := &service.InfoNode{Type: "metadata", NodeType: service.NodeTypeInfoNode}
+	tmpInfoNode := &service.InfoNode{Type: "metadata"}
 	for key, val := range phase.masterConfig.MetaData {
-		tmpInfoNode.DataNodes = append(tmpInfoNode.DataNodes, &service.InfoNode_DataNode{Type: key, Data: val, NodeType: service.NodeTypeDataNode})
+		tmpInfoNode.DataNodes = append(tmpInfoNode.DataNodes, &service.InfoNode_DataNode{Type: key, Data: val})
 	}
 
 	if len(tmpInfoNode.DataNodes) > 0 {
