+++ proto/datamodel.proto
@@ -4,7 +4,7 @@ option java_package = "org.qmstr.grpc.service";
 
 message FileNode {
     string uid = 1;
-    int32 nodeType =2;
+    string fileNodeType =2;
     string type = 3;
     string path = 4;
     string name = 5;
@@ -16,13 +16,13 @@ message FileNode {
 
 message InfoNode {
     string uid = 1;
-    int32 nodeType = 2;
+    string infoNodeType = 2;
     string type = 3;
     double confidenceScore = 4;
     repeated Analyzer analyzer = 5;
 
     message DataNode {
-        int32 nodeType = 1;
+        string dataNodeType = 1;
         string type = 2;
         string data = 3;
     }
@@ -32,7 +32,7 @@ message InfoNode {
 message Analyzer {
     string uid = 1;
     string name = 2;
-    int32 nodeType = 3;
+    string analyzerNodeType = 3;
     int64 trustLevel = 4;
     repeated PathSubstitution pathSub = 5;
 }
@@ -46,7 +46,7 @@ message PackageNode {
     string uid = 1;
     string name = 2;
     string session = 3;
-    int32 nodeType = 4;
+    string packageNodeType = 4;
     repeated FileNode targets = 5;
     repeated InfoNode additionalInfo = 6;
     string buildConfig = 7;
